# api/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        example: session not found
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Web-based KVM panel for managing remote server consoles via BMC/IPMI.
    Supports AMI MegaRAC (native iKVM), Dell iDRAC (VNC/WebSocket), and NanoKVM devices.
    Optional OIDC authentication with role-based access control.
  title: KVM Switcher API
  version: "1.0"
paths:
...
  /ws/kvm/{id}:
    get:
      description: Upgrades to a WebSocket connection and proxies VNC/iKVM traffic
        between the browser and the BMC. Uses the "binary" WebSocket subprotocol.
        For iKVM sessions, the bridge authenticates with the BMC and establishes an
        IVTP tunnel. For WebSocket sessions, proxies to the remote WSS endpoint. For
        VNC sessions, proxies to the raw TCP VNC port.
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "101":
          description: WebSocket upgrade successful (binary subprotocol)
        "404":
          description: Session not found or not connected
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: KVM WebSocket proxy
      tags:
      - websocket
securityDefinitions:
  CookieAuth:
    in: cookie
    name: kvm_session
    type: apiKey
swagger: "2.0"